+++ packages/graphqlgen/benchmarks/integration/index.ts
@@ -1,20 +1,15 @@
-/**
- * This module contains utility functions for creating
- * integration benchmark instances.
- */
-
+import * as Path from 'path'
+import * as Glob from 'glob'
 import * as Parse from '../../src/parse'
 import * as ConfigTypes from 'graphqlgen-json-schema'
 import * as Validation from '../../src/validation'
 import * as GGen from '../../src'
-import * as Path from 'path'
-import * as Glob from 'glob'
 import * as Util from '../../src/utils'
-import { Benchmark } from '../lib/benchmark'
+import * as Benchmark from '../lib/benchmark'
 
-const collect = (): Benchmark[] => {
+const collect: Benchmark.Collect = () => {
   const paths = Glob.sync(Path.join(__dirname, './*'))
-  const benchmarks: Benchmark[] = []
+  const benchmarks: Benchmark.Benchmark[] = []
 
   for (const path of paths) {
     if (Util.isFile(path)) continue
@@ -32,8 +27,9 @@ const collect = (): Benchmark[] => {
 
     for (const modelPath of modelPaths) {
       // 1. We know there will be an extension because of isFile
-      // filter above. 2. We know it will be a support language extension
-      // because of the filter right after.
+      //    filter above.
+      // 2. We know it will be a support language extension
+      //    because of the set filter below.
       const ext = Util.getExt(modelPath) as Util.LanguageExtension
 
       if (!Util.languageExtensions.includes(ext)) continue
@@ -60,13 +56,13 @@ type Options = {
  * of the whole GraphqlGen pipeline (except for initial
  * config parsing, file loading, and model map creation).
  */
-const create = (config: Options): Benchmark => {
+const create = (config: Options): Benchmark.Benchmark => {
   const codeGenConfig = createCodeGenConfig({
     language: config.language,
     rootPath: config.rootPath,
   })
 
-  const benchmark = new Benchmark({
+  const benchmark = new Benchmark.Benchmark({
     name: `generateCode (${Path.basename(config.rootPath)} schema, ${
       config.language
     })`,
