+++ packages/graphqlgen/benchmarks/integration/trivial/models.ts
