+++ packages/graphqlgen/benchmarks/integration/trivial/schema.graphql
