+++ packages/graphqlgen/benchmarks/lib/benchmark.ts
@@ -1,5 +1,11 @@
 import * as Bench from 'benchmark'
+import * as H from '../lib/helpers'
 
+type History = Record<string, Report[]>
+
+/**
+ * The information about how a benchmark performed.
+ */
 type Report = {
   name: string
   summary: string
@@ -8,6 +14,11 @@ type Report = {
   cycles: number
   stats: Bench.Stats
 }
+
+/**
+ * Function that will collect benchmarks of a benchmark type.
+ */
+type Collect = () => Benchmark[]
 
 class Benchmark {
   runner: Bench
@@ -34,4 +45,15 @@ class Benchmark {
   }
 }
 
-export { Benchmark, Report }
+/**
+ * Save results of benchmarks to a json file. Provided path must point to a
+ * JSON file. File must at least contain the seed.
+ */
+const saveReports = (path: string, reports: Report[]): void => {
+  H.updateJSONFile<History>(path, history => {
+    history[H.getGitHeadSha()] = reports
+    return history
+  })
+}
+
+export { Report, Benchmark, saveReports, Collect }
