+++ packages/graphqlgen/benchmarks/main.ts
@@ -1,32 +1,7 @@
-import * as IntegrationBenchmarks from './integration'
 import * as Benchmark from './lib/benchmark'
+import * as IntegrationBenchmarks from './integration'
 import * as yargs from 'yargs'
-import * as FS from 'fs'
 import * as Path from 'path'
-import * as CP from 'child_process'
-
-type History = Record<string, Benchmark.Report[]>
-
-const printReports = (reports: Benchmark.Report[]): void => {
-  for (const report of reports) {
-    console.log(report.summary)
-  }
-}
-
-const getGitHeadSha = (): string => {
-  return CP.execSync('git rev-parse HEAD', { encoding: 'utf8' }).trim()
-}
-
-const updateJSONFile = <T extends unknown>(
-  path: string,
-  updater: (data: T) => T,
-): void => {
-  const data: T = JSON.parse(
-    FS.readFileSync(Path.join(__dirname, './history.json'), 'utf8'),
-  )
-  const updatedData = updater(data)
-  FS.writeFileSync(path, JSON.stringify(updatedData, null, 2))
-}
 
 const argv = yargs
   .usage('bench [args]')
@@ -47,11 +22,8 @@ const reports = IntegrationBenchmarks.collect().reduce<Benchmark.Report[]>(
 )
 
 if (argv.save) {
-  updateJSONFile<History>(Path.join(__dirname, './history.json'), history => {
-    history[getGitHeadSha()] = reports
-    return history
-  })
-  printReports(reports)
+  Benchmark.saveReports(Path.join(__dirname, './history.json'), reports)
+  reports.forEach(report => console.log(report.summary))
 } else {
-  printReports(reports)
+  reports.forEach(report => console.log(report.summary))
 }
